requirements.txt
python-telegram-bot==20.7
firebase-admin==6.4.0
python-dotenv==1.0.0
flask
telegram
firebase-admin==6.4.0
flask
python-dotenv==1.0.0
python-telegram-bot==20.7
telegram